+++ PixelDash/HUD.cpp
@@ -35,25 +35,25 @@ void HUD::init(olc::PixelGameEngine& pge)
 	if (!instance) {
 		instance = new HUD();
 		instance->pge = &pge;
-		instance->liveBar = new PixelSprite(pge, "assets/LiveBar.png", 1, 0.1, 66, 34, 4, 0);
+		instance->liveBar = new PixelSprite(pge, "assets/LiveBar.png", 1, 0.1f, 66, 34, 4, 0);
 		instance->liveBar->setPosY(1);
 
-		instance->smallDiamond = new PixelSprite(pge, "assets/SmallDiamond.png", 7, 0.1, 18, 14, 0, 0);
+		instance->smallDiamond = new PixelSprite(pge, "assets/SmallDiamond.png", 7, 0.1f, 18, 14, 0, 0);
 		instance->smallDiamond->setLoop(true);
 		instance->smallDiamond->setAnimation(true);
 		instance->smallDiamond->setPosX(11);
 		instance->smallDiamond->setPosY(25);
 
-		instance->numberOne = new PixelSprite(pge, "assets/Numbers.png", 30, 27, 1, 0.1, 6, 8, 54, 0, 0, 0);
-		instance->numberTwo = new PixelSprite(pge, "assets/Numbers.png", 35, 27, 1, 0.1, 6, 8, 54, 0, 0, 0);
+		instance->numberOne = new PixelSprite(pge, "assets/Numbers.png", 30, 27, 1, 0.1f, 6, 8, 54, 0, 0, 0);
+		instance->numberTwo = new PixelSprite(pge, "assets/Numbers.png", 35, 27, 1, 0.1f, 6, 8, 54, 0, 0, 0);
 
 		for (int i = 0; i < instance->m_nTotalLife; ++i) {
 			int offset = 11;
-			PixelSprite* heart = new PixelSprite(pge, "assets/SmallHeart.png", 7, 0.1, 18, 14, 4, 0);
+			PixelSprite* heart = new PixelSprite(pge, "assets/SmallHeart.png", 7, 0.1f, 18, 14, 4, 0);
 			heart->setLoop(true);
 			heart->setAnimation(true);
-			heart->setPosX(offset * (i + 1));
-			heart->setPosY(offset);
+			heart->setPosX((float)(offset * (i + 1)));
+			heart->setPosY((float)offset);
 
 			instance->lifeHearts.emplace_back(heart);
 		}
