+++ PixelDash/Platform.h
@@ -15,7 +15,7 @@ enum class PlatformType {
 class Platform : public PixelSprite {
 public:
     Platform(olc::PixelGameEngine& pge, PlatformType type, float fposx = 0, float fposy = 0)
-        : PixelSprite(pge, "assets/Decorations32x32.png", 0, 0, 1, 0.1, 32, 40, getOffsetX(type), getOffsetY(type), 0, 0)
+        : PixelSprite(pge, "assets/Decorations32x32.png", 0, 0, 1, 0.1f, 32, 40, (int)getOffsetX(type), (int)getOffsetY(type), 0, 0)
     {
         setSprOffsetX(getOffsetX(type));
         setSprOffsetY(getOffsetY(type));
@@ -29,18 +29,18 @@ public:
 private:
     float homeX;
     float homeY;
-    static int getOffsetX(PlatformType type) {
+    static float getOffsetX(PlatformType type) {
         switch (type) {
-        case PlatformType::ONE:   return 64;
-        case PlatformType::TWO:   return 96;
-        case PlatformType::THREE: return 128;
-        case PlatformType::FOUR:  return 160;
+        case PlatformType::ONE:   return 64.0f;
+        case PlatformType::TWO:   return 96.0f;
+        case PlatformType::THREE: return 128.0f;
+        case PlatformType::FOUR:  return 160.0f;
         }
-        return 0; // Default case
+        return 0.0f; // Default case
     }
 
-    static int getOffsetY(PlatformType type) {
-        return 40;  // If all platforms share the same Y-offset, otherwise add logic here
+    static float getOffsetY(PlatformType type) {
+        return 40.0f;  // If all platforms share the same Y-offset, otherwise add logic here
     }
 };
